Improve partial inference of key container operations
`resize!` and `sizehint!` are big targets for invalidation, and quite
a lot of code was calling these methods with poor inference about the
size. Since the size is just `Int`, asserting that eliminates numerous
vulnerabilities. Likewise, bounds-checking always works with `Bool`.
Finally, add a couple of short-circuits for `eltype(::Tuple{T,...})`,
including one that bypasses `typejoin`.

+++ base/abstractset.jl
@@ -86,7 +86,7 @@ max_values(::Type{Bool}) = 2
 max_values(::Type{Nothing}) = 1
 
 function union!(s::AbstractSet{T}, itr) where T
-    haslength(itr) && sizehint!(s, length(s) + length(itr))
+    haslength(itr) && sizehint!(s, length(s) + Int(length(itr))::Int)
     for x in itr
         push!(s, x)
         length(s) == max_values(T) && break
