+++ base/dict.jl
@@ -99,7 +99,7 @@ mutable struct Dict{K,V} <: AbstractDict{K,V}
 end
 function Dict{K,V}(kv) where V where K
     h = Dict{K,V}()
-    haslength(kv) && sizehint!(h, length(kv))
+    haslength(kv) && sizehint!(h, Int(length(kv))::Int)
     for (k,v) in kv
         h[k] = v
     end
@@ -166,7 +166,7 @@ end
 
 empty(a::AbstractDict, ::Type{K}, ::Type{V}) where {K, V} = Dict{K, V}()
 
-hashindex(key, sz) = (((hash(key)%Int) & (sz-1)) + 1)::Int
+hashindex(key, sz) = (((hash(key)::UInt % Int) & (sz-1)) + 1)::Int
 
 @propagate_inbounds isslotempty(h::Dict, i::Int) = h.slots[i] == 0x0
 @propagate_inbounds isslotfilled(h::Dict, i::Int) = h.slots[i] == 0x1
@@ -239,7 +239,7 @@ function sizehint!(d::Dict{T}, newsz) where T
     end
     # grow at least 25%
     newsz = min(max(newsz, (oldsz*5)>>2),
-                max_values(T))
+                max_values(T)::Int)
     rehash!(d, newsz)
 end
